MB-32461: Change two config vars representing seconds to float
Changing:

* bucket_quota_change_task_poll_interval
* ht_resize_interval

Both can be redefined as float which will allow for sub-second snoozing
and maybe useful in stress testing.

As these parameters are only consumed by direct calls to
GlobalTask::snooze they can be switched to float.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228045
Tested-by: Jim Walker <jim@couchbase.com>

+++ engines/ep/src/ep_engine.cc
@@ -675,7 +675,7 @@ cb::engine_errc EventuallyPersistentEngine::setFlushParam(
         } else if (key == "getl_max_timeout") {
             configuration.setGetlMaxTimeout(std::stoull(val));
         } else if (key == "ht_resize_interval") {
-            configuration.setHtResizeInterval(std::stoull(val));
+            configuration.setHtResizeInterval(std::stof(val));
         } else if (key == "ht_resize_algo") {
             configuration.setHtResizeAlgo(val);
         } else if (key == "ht_size") {
@@ -855,8 +855,7 @@ cb::engine_errc EventuallyPersistentEngine::setFlushParam(
         } else if (key == "range_scan_max_continue_tasks") {
             configuration.setRangeScanMaxContinueTasks(std::stoul(val));
         } else if (key == "bucket_quota_change_task_poll_interval") {
-            configuration.setBucketQuotaChangeTaskPollInterval(std::stoul
-                                                                (val));
+            configuration.setBucketQuotaChangeTaskPollInterval(std::stof(val));
         } else if (key == "range_scan_read_buffer_send_size") {
             configuration.setRangeScanReadBufferSendSize(std::stoull(val));
         } else if (key == "range_scan_max_lifetime") {
